Move timeTravel query logic to service 🎨
