+++ server/revision/revision.controller.js
@@ -1,10 +1,9 @@
 'use strict';
 
-const { Activity, Revision, Sequelize, User } = require('../shared/database');
+const { Activity, Revision, User } = require('../shared/database');
+const { getEntityRemovesSinceMoment, getLastState } = require('./revision.service.js');
 const map = require('lodash/map');
 const { resolveStatics } = require('../shared/storage/helpers');
-
-const { Op } = Sequelize;
 
 function index({ repository, query, opts }, res) {
   const { entity, entityId } = query;
@@ -28,7 +27,7 @@ async function getStateAtMoment({ query }, res) {
   const removes = await getEntityRemovesSinceMoment(activity, timestamp);
   const entityIds = [...elementIds, ...map(removes.elements, 'state.id')];
   const removedActivityIds = map(removes.activities, 'state.id');
-  const elements = await getLastRevision(entityIds, removedActivityIds, timestamp);
+  const elements = await getLastState(entityIds, removedActivityIds, timestamp);
   return res.json({ data: { ...removes, elements } });
 }
 
@@ -44,52 +43,3 @@ module.exports = {
   getStateAtMoment,
   resolve
 };
-
-async function getEntityRemovesSinceMoment(activity, timestamp) {
-  const { nodes } = await activity.descendants({ paranoid: false });
-  const whereRemovedAfter = {
-    operation: 'REMOVE',
-    createdAt: { [Op.gt]: timestamp }
-  };
-  const whereCreatedBefore = {
-    createdAt: { [Op.lt]: timestamp }
-  };
-  const hasNodeId = { [Op.in]: map(nodes, 'id') };
-  const [activities, elements] = await Promise.all([
-    Revision.findAll({
-      where: {
-        ...whereRemovedAfter,
-        entity: 'ACTIVITY',
-        state: { ...whereCreatedBefore, id: hasNodeId }
-      }
-    }),
-    Revision.findAll({
-      where: {
-        ...whereRemovedAfter,
-        entity: 'CONTENT_ELEMENT',
-        state: { ...whereCreatedBefore, activityId: hasNodeId }
-      }
-    })
-  ]);
-  return { activities, elements };
-}
-
-function getLastRevision(ids, activityIds, beforeTimestamp) {
-  const whereCreateOrUpdate = {
-    operation: { [Op.or]: ['CREATE', 'UPDATE'] }
-  };
-  const whereBefore = { createdAt: { [Op.lt]: beforeTimestamp } };
-  return Revision.scope('lastByEntity').findAll({
-    where: {
-      ...whereCreateOrUpdate,
-      ...whereBefore,
-      state: {
-        [Op.or]: [{
-          id: { [Op.in]: ids }
-        }, {
-          activityId: { [Op.in]: activityIds }
-        }]
-      }
-    }
-  });
-}
